Return money of the last transaction when canceling tower placement
Fixed #12.
Note that to prevent being able to place several towers for free, I made it so that after having placed one tower, the cursor behaviour goes back to placeTower = false.
We might want to make it possible to place several towers of the same type in one run, but then we need to think about the budget handling.