Add first three tower types using scene inheritance
Currently only the texture changes and thus tower1 is the same as the base tower scene, but this might change later on as we make tower types more complex.
Had to revert the tower scenes to xscn format though due to a current upstream bug with tscn and scene inheritance.